feat(ibm): create shared tf configs for team tests
