+++ sysdig/resource_sysdig_monitor_team_test.go
@@ -47,31 +47,6 @@ func TestAccMonitorTeam(t *testing.T) {
 	})
 }
 
-func monitorTeamMinimumConfiguration(name string) string {
-	return fmt.Sprintf(`
-resource "sysdig_monitor_team" "sample" {
-  name      = "sample-%s"
-
-  entrypoint {
-	type = "Explore"
-  }
-}`, name)
-}
-
-func monitorTeamWithName(name string) string {
-	return fmt.Sprintf(`
-resource "sysdig_monitor_team" "sample" {
-  name               = "sample-%s"
-  description        = "%s"
-  scope_by           = "container"
-  filter             = "container.image.repo = \"sysdig/agent\""
-
-  entrypoint {
-	type = "Explore"
-  }
-}`, name, name)
-}
-
 func monitorTeamWithFullConfig(name string) string {
 	return fmt.Sprintf(`
 resource "sysdig_monitor_team" "sample" {
